completed implementation of all method in TransactionDataFetcher except for last one.

+++ coding_test/src/main/java/com/smallworld/model/Transaction.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class Transaction {
 
-    private Long mtn;
+    private Integer mtn;
     private Double amount;
     private String senderFullName;
     private Integer senderAge;
